+++ Cargo.toml
@@ -102,47 +102,25 @@ lto = true
 codegen-units = 1
 
 [package.metadata.leptos]
-# The name used by wasm-bindgen/cargo-leptos for the JS/WASM bundle. Defaults to the crate name   
 output-name = "lemmy-ui-leptos"
-# The site root folder is where cargo-leptos generate all output. WARNING: all content of this folder will be erased on a rebuild. Use it in your server setup.
-site-root = "target/site"
-# The site-root relative folder where all compiled output (JS, WASM and CSS) is written
-# Defaults to pkg	
-site-pkg-dir = "pkg"
+
 tailwind-input-file = "style/tailwind.css"
-# [Optional] Files in the asset-dir will be copied to the site-root directory
+browserquery = "defaults"
+
+site-root = "target/site"
+site-pkg-dir = "pkg"
 assets-dir = "public"
-# The IP and port (ex: 127.0.0.1:3000) where the server serves the content. Use it in your server setup.
 site-addr = "0.0.0.0:1237"
-# The port to use for automatic reload monitoring
 reload-port = 3001
-# [Optional] Command to use when running end2end tests. It will run in the end2end dir.
+
 end2end-cmd = "npx playwright test"
 end2end-dir = "end2end"
-#  The browserlist query used for optimizing the CSS.
-browserquery = "defaults"
-# Set by cargo-leptos watch when building with tha tool. Controls whether autoreload JS will be included in the head
+
 watch = false
-# The environment Leptos will run in, usually either "DEV" or "PROD"
-env = "DEV"
-# The features to use when compiling the bin target
-#
-# Optional. Can be over-ridden with the command line parameter --bin-features
-bin-features = ["ssr", "bypass_internal_proxy"]
-
-# If the --no-default-features flag should be used when compiling the bin target
-#
-# Optional. Defaults to false.
+bin-features = ["ssr"]
 bin-default-features = false
-
-# The features to use when compiling the lib target
-#
-# Optional. Can be over-ridden with the command line parameter --lib-features
-lib-features = ["hydrate", "bypass_internal_proxy"]
-
-# If the --no-default-features flag should be used when compiling the lib target
-#
-# Optional. Defaults to false.
+lib-features = ["hydrate"]
 lib-default-features = false
 
+env = "DEV"
 lib-profile-release = "wasm-release"
