--- src/server/mod.rs
@@ -1,6 +1,0 @@
-#[cfg(all(feature = "ssr", not(feature = "bypass_internal_proxy")))]
-mod api_service;
-
-#[cfg(feature = "ssr")]
-mod cookie_middleware;
-
