+++ src/ui/components/common/nav.rs
@@ -57,9 +57,9 @@ pub fn TopNav() -> impl IntoView {
         <div class="navbar-start">
           <ul class="menu menu-horizontal flex-nowrap">
             <li>
-              <A href="/" class="text-xl whitespace-nowrap">
-                {instance_name}
-              </A>
+              // <A href="/" class="text-xl whitespace-nowrap">
+              //   {instance_name}
+              // </A>
             </li>
             <li>
               <A href="/communities" class="text-md">
@@ -110,55 +110,55 @@ pub fn TopNav() -> impl IntoView {
                 </ul>
               </details>
             </li>
-            <Show
-              when=move || with!(| my_user | my_user.is_some())
-              fallback=move || {
-                  view! {
+            // <Show
+            //   when=move || with!(| my_user | my_user.is_some())
+            //   fallback=move || {
+            //       view! {
                     <li>
                       <A href="/login">{t!(i18n, login)}</A>
                     </li>
                     <li>
                       <A href="/signup">{t!(i18n, signup)}</A>
                     </li>
-                  }
-              }
-            >
+            //       }
+            //   }
+            // >
 
-              <li>
-                <A href="/inbox">
-                  <span title=t!(i18n, unread_messages)>
-                    <Icon icon=Icon::from(ChIcon::ChBell) class="h-6 w-6"/>
-                  </span>
-                </A>
-              </li>
-              <li>
-                <details>
-                  <summary>
-                    {with!(
-                        | my_user | { let Person { name, display_name, .. } = my_user.as_ref()
-                        .unwrap(); display_name.as_ref().unwrap_or(name).to_string() }
-                    )}
+            //   <li>
+            //     <A href="/inbox">
+            //       <span title=t!(i18n, unread_messages)>
+            //         <Icon icon=Icon::from(ChIcon::ChBell) class="h-6 w-6"/>
+            //       </span>
+            //     </A>
+            //   </li>
+            //   <li>
+            //     <details>
+            //       <summary>
+            //         {with!(
+            //             | my_user | { let Person { name, display_name, .. } = my_user.as_ref()
+            //             .unwrap(); display_name.as_ref().unwrap_or(name).to_string() }
+            //         )}
 
-                  </summary>
-                  <ul class="z-10">
-                    <li>
-                      <A href=with!(
-                          | my_user | format!("/u/{}", my_user.as_ref().unwrap().name)
-                      )>{t!(i18n, profile)}</A>
-                    </li>
-                    <li>
-                      <A href="/settings">{t!(i18n, settings)}</A>
-                    </li>
-                    <div class="divider my-0"></div>
-                    <li>
-                      <ActionForm action=logout_action>
-                        <button type="submit">{t!(i18n, logout)}</button>
-                      </ActionForm>
-                    </li>
-                  </ul>
-                </details>
-              </li>
-            </Show>
+            //       </summary>
+            //       <ul class="z-10">
+            //         <li>
+            //           <A href=with!(
+            //               | my_user | format!("/u/{}", my_user.as_ref().unwrap().name)
+            //           )>{t!(i18n, profile)}</A>
+            //         </li>
+            //         <li>
+            //           <A href="/settings">{t!(i18n, settings)}</A>
+            //         </li>
+            //         <div class="divider my-0"></div>
+            //         <li>
+            //           <ActionForm action=logout_action>
+            //             <button type="submit">{t!(i18n, logout)}</button>
+            //           </ActionForm>
+            //         </li>
+            //       </ul>
+            //     </details>
+            //   </li>
+            // </Show>
           </ul>
         </div>
       </nav>
